Move some functions to comment.py
This required quite some re-plumbing.
