+++ src/reuse/report.py
@@ -32,10 +32,11 @@ from typing import (
 from uuid import uuid4
 
 from . import __REUSE_version__, __version__
-from ._util import _LICENSEREF_PATTERN, _LICENSING, StrPath, _checksum
+from ._util import _LICENSEREF_PATTERN, _LICENSING, _checksum
 from .global_licensing import ReuseDep5
 from .i18n import _
 from .project import Project, ReuseInfo
+from .types import StrPath
 
 _LOGGER = logging.getLogger(__name__)
